refactor: change file naming for conflict prevention (#5)
